Add files via upload
The objective of this project is to design and implement a D Flip-Flop circuit using the Microwind tool, a digital circuit design and simulation software. The D Flip-Flop is a fundamental digital circuit element that is widely used in digital systems for storing and transferring data.The project will be graded based on the Correctness and completeness of the circuit design, Accuracy of the simulation results.